Start on Android implementatio of Wallet
Android has certain apps that can handle your Apple Wallet Passes. I
found out that the approach I took in this commit, will never work, as I
misunderstood what the shoebox urlscheme does on iOS. It will open Apple
Pay, it's not something they use in PassKit to add passes. I figured
this out when I installed a few of those Android apps and they didn't
really care about my urlscheme. When I removed the code that changes the
uri's scheme to shoebox and the browser offered me to download the
pkpass file for me. All the Android apps were able to add the pkpass.

Next step will be to try and download the file in a Downloads folder and
offer an intent to any app that says it can handle the file.

Re #3